flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/space/DeleteSpaceService.java
package org.flickit.assessment.users.application.service.space;

import lombok.RequiredArgsConstructor;
import org.flickit.assessment.common.exception.AccessDeniedException;
import org.flickit.assessment.users.application.port.in.space.DeleteSpaceUseCase;
import org.flickit.assessment.users.application.port.out.space.LoadSpaceOwnerPort;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Objects;
import java.util.UUID;

import static org.flickit.assessment.common.error.ErrorMessageKey.COMMON_CURRENT_USER_NOT_ALLOWED;

@Service
@Transactional
@RequiredArgsConstructor
public class DeleteSpaceService implements DeleteSpaceUseCase {

    private final LoadSpaceOwnerPort loadSpaceOwnerPort;

    @Override
    public void deleteSpace(Param param) {
        validateCurrentUser(param.getId(), param.getCurrentUserId());
    }

    private void validateCurrentUser(Long spaceId, UUID currentUserId) {
        UUID expertGroupOwnerId = loadSpaceOwnerPort.loadOwnerId(spaceId);
        if (!Objects.equals(expertGroupOwnerId, currentUserId))
            throw new AccessDeniedException(COMMON_CURRENT_USER_NOT_ALLOWED);
    }
}
